Add files via upload
A landing page in web development is a standalone web page, created specifically for marketing or advertising purposes. It's where visitors "land" after clicking on a link in an email, social media post, or search engine result. The primary goal of a landing page is to convert visitors into leads or customers through a specific call-to-action (CTA), such as signing up for a newsletter, downloading a resource, or purchasing a product/service.